Introduce ClippingRenderContext
