+++ Source/OxyPlot/Rendering/RenderContext/RenderContextBase.cs
@@ -17,8 +17,6 @@ namespace OxyPlot
     /// </summary>
     public abstract class RenderContextBase : IRenderContext
     {
-        private readonly Stack<OxyRect> clipStack = new Stack<OxyRect>();
-
         /// <summary>
         /// Initializes a new instance of the <see cref="RenderContextBase" /> class.
         /// </summary>
@@ -231,42 +229,10 @@ namespace OxyPlot
         }
 
         /// <inheritdoc/>
-        public virtual void PopClip()
-        {
-            if (this.clipStack.Count == 0)
-            {
-                throw new InvalidOperationException($"Unbalanced call to {nameof(PopClip)}.");
-            }
-
-            this.ResetClip();
-            this.clipStack.Pop();
-            if (this.clipStack.Count > 0)
-            {
-                this.SetClip(this.clipStack.Peek());
-            }
-        }
+        public abstract void PopClip();
 
         /// <inheritdoc/>
-        public virtual void PushClip(OxyRect clippingRectangle)
-        {
-            if (this.clipStack.Count > 0)
-            {
-                var currentClippingRectangle = this.clipStack.Peek();
-                var newClippingRectangle = clippingRectangle.Intersect(currentClippingRectangle);
-                if (!currentClippingRectangle.Equals(newClippingRectangle))
-                {
-                    this.ResetClip();
-                    this.SetClip(newClippingRectangle);
-                }
-
-                this.clipStack.Push(newClippingRectangle);
-            }
-            else
-            {
-                this.SetClip(clippingRectangle);
-                this.clipStack.Push(clippingRectangle);
-            }
-        }
+        public abstract void PushClip(OxyRect clippingRectangle);
 
         /// <summary>
         /// Creates an ellipse polygon.
@@ -356,19 +322,5 @@ namespace OxyPlot
                     return true;
             }
         }
-
-        /// <summary>
-        /// Sets the clipping area to the specified rectangle.
-        /// </summary>
-        /// <param name="clippingRectangle">The rectangle.</param>
-        /// <remarks>
-        /// Calls to this functions must always be balanced by a call to <see cref="ResetClip"/> before calling <see cref="SetClip(OxyRect)"/> again.
-        /// </remarks>
-        protected abstract void SetClip(OxyRect clippingRectangle);
-
-        /// <summary>
-        /// Resets the clipping area.
-        /// </summary>
-        protected abstract void ResetClip();
     }
 }
